Refactor and merge `QueryDataReader` and `AbstractQueryDataReader` to single class
