+++ tests/Support/CustomerDataReader.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Yiisoft\Data\Db\Tests\Support;
 
-use Yiisoft\Data\Db\AbstractQueryDataReader;
+use Yiisoft\Data\Db\QueryDataReader;
 
-final class CustomerDataReader extends AbstractQueryDataReader
+final class CustomerDataReader extends QueryDataReader
 {
     protected function createItem(object|array $row): array|object
     {
